backend/requirements.txt
alipay-sdk-python==3.7.249
asgiref==3.8.1
boto3==1.36.26
botocore==1.36.26
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-ratelimit==4.1.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
optionaldict==0.1.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycryptodome==3.20.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
s3transfer==0.11.2
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
wheel==0.44.0
whitenoise==6.8.2
xmltodict==0.13.0
